Extract Graph Cover
